Change return type of add function in add.cpp from double to int for consistency

+++ 01-basics/Practicing Functions/Multifiles header/add.cpp
@@ -2,7 +2,7 @@
 #include "add.h"
 #include <iostream>
 
-double add(int x, int y)
+int add(int x, int y)
 {
     return x+y;
 }
